FunSQL.jl

API Reference

SQL Dialects

FunSQL.SQLDialect — Type

Properties of a SQL dialect.

SQL Entities

FunSQL.SQLTable — Type

SQLTable(; schema = nothing, name, columns)
SQLTable(name; schema = nothing, columns)
SQLTable(name, columns...; schema = nothing)

The structure of a SQL table or a table-like entity (TEMP TABLE, VIEW, etc) for use as a reference in assembling SQL queries.

The SQLTable constructor expects the table name, a vector columns of column names, and, optionally, the name of the table schema. A name can be provided as a Symbol or String value.

Examples

julia> t = SQLTable(:location,
                    :location_id, :address_1, :address_2, :city, :state, :zip);

julia> show(t.name)
:location

julia> show(t.columns)
[:location_id, :address_1, :address_2, :city, :state, :zip]

julia> t = SQLTable(schema = "public",
                    name = "person",
                    columns = ["person_id", "birth_datetime", "location_id"]);

julia> show(t.schema)
:public

julia> show(t.name)
:person

julia> show(t.columns)
[:person_id, :birth_datetime, :location_id]

Syntactic Structure

FunSQL.AbstractSQLClause — Type

A part of a SQL query.

FunSQL.SQLClause — Type

An opaque wrapper over an arbitrary SQL clause.

FunSQL.render — Method

render(clause; dialect = :default) :: String

Convert the given SQL clause object to a SQL string.

FunSQL.AS — Method

AS(; over = nothing, name)
AS(name; over = nothing)

An AS clause.

Examples

julia> c = ID(:person) |> AS(:p);

julia> print(render(c))
"person" AS "p"

FunSQL.FROM — Method

FROM(; over = nothing)
FROM(over)

A FROM clause.

Examples

julia> c = ID(:person) |> AS(:p) |> FROM();

julia> print(render(c))
FROM "person" AS "p"

FunSQL.ID — Method

ID(; over = nothing, name)
ID(name; over = nothing)

A SQL identifier. Specify over or use the |> operator to make a qualified identifier.

Examples

julia> c = ID(:person);

julia> print(render(c))
"person"

julia> c = ID(:p) |> ID(:birth_datetime);

julia> print(render(c))
"p"."birth_datetime"

FunSQL.LITERAL — Method

LITERAL(val)

A SQL literal.

In a context of a SQL clause, missing, numbers, strings and datetime values are automatically converted to SQL literals.

Examples

julia> c = LITERAL(missing);

julia> print(render(c))
NULL

julia> c = LITERAL("SQL is fun!");

julia> print(render(c))
'SQL is fun!'

FunSQL.SELECT — Method

SELECT(; over = nothing, distinct = false, list)
SELECT(list...; over = nothing, distinct = false)

A SELECT clause. Unlike raw SQL, SELECT() should be placed at the end of a clause chain.

Set distinct to true to add a DISTINCT modifier.

Examples

julia> c = SELECT(true, false);

julia> print(render(c))
SELECT TRUE, FALSE

julia> c = FROM(:location) |>
           SELECT(distinct = true, :zip);

julia> print(render(c))
SELECT DISTINCT zip
FROM location

FunSQL.WHERE — Method

WHERE(; over = nothing, condition)
WHERE(condition; over = nothing)

A WHERE clause.

Examples

julia> c = FROM(:location) |>
           WHERE(OP("=", :zip, "60614")) |>
           SELECT(:location_id);

julia> print(render(c))
SELECT "location_id"
FROM "location"
WHERE "zip" = '60614'