API Reference

FunSQL.render — Function

render(node; dialect = :default) :: String

Convert the given SQL node or clause object to a SQL string.

SQL Dialects

FunSQL.SQLDialect — Type

Properties of a SQL dialect.

SQL Entities

FunSQL.SQLTable — Type

SQLTable(; schema = nothing, name, columns)
SQLTable(name; schema = nothing, columns)
SQLTable(name, columns...; schema = nothing)

The structure of a SQL table or a table-like entity (TEMP TABLE, VIEW, etc) for use as a reference in assembling SQL queries.

The SQLTable constructor expects the table name, a vector columns of column names, and, optionally, the name of the table schema. A name can be provided as a Symbol or String value.

Examples

julia> t = SQLTable(:location,
                    :location_id, :address_1, :address_2, :city, :state, :zip);

julia> show(t.name)
:location

julia> show(t.columns)
[:location_id, :address_1, :address_2, :city, :state, :zip]

julia> t = SQLTable(schema = "public",
                    name = "person",
                    columns = ["person_id", "birth_datetime", "location_id"]);

julia> show(t.schema)
:public

julia> show(t.name)
:person

julia> show(t.columns)
[:person_id, :birth_datetime, :location_id]

SQL Statements

FunSQL.SQLStatement — Type

Prepared SQL statement.

FunSQL.pack — Function

pack(stmt::SQLStatement, vars::Union{Dict, NamedTuple}) :: Vector{Any}

Convert named parameters to positional form.

Semantic Structure

FunSQL.AbstractSQLNode — Type

A SQL expression.

FunSQL.SQLNode — Type

An opaque wrapper over an arbitrary SQL node.

FunSQL.Agg — Method

Agg(; over = nothing, name, distinct = false, args = [], filter = nothing)
Agg(name; over = nothing, distinct = false, args = [], filter = nothing)
Agg(name, args...; over = nothing, distinct = false, filter = nothing)

An application of an aggregate function.

Example

julia> person = SQLTable(:person, columns = [:person_id, :year_of_birth]);

julia> q = From(person) |>
           Group(Get.year_of_birth) |>
           Select(Get.year_of_birth, Agg.count());

julia> print(render(q))
SELECT "person_1"."year_of_birth", COUNT(*) AS "count"
FROM "person" AS "person_1"
GROUP BY "person_1"."year_of_birth"

julia> person = SQLTable(:person, columns = [:person_id, :year_of_birth]);

julia> q = From(person) |>
           Group() |>
           Select(Agg.count(distinct = true, Get.year_of_birth));

julia> print(render(q))
SELECT COUNT(DISTINCT "person_1"."year_of_birth") AS "count"
FROM "person" AS "person_1"

julia> person = SQLTable(:person, columns = [:person_id]);

julia> visit_occurrence =
           SQLTable(:visit_occurrence, columns = [:visit_occurrence_id, :person_id, :visit_start_date]);

julia> q = From(person) |>
           LeftJoin(:visit_group => From(visit_occurrence) |>
                                    Group(Get.person_id),
                    on = (Get.person_id .== Get.visit_group.person_id)) |>
           Select(Get.person_id,
                  :max_visit_start_date =>
                      Get.visit_group |> Agg.max(Get.visit_start_date));

julia> print(render(q))
SELECT "person_1"."person_id", "visit_group_1"."max" AS "max_visit_start_date"
FROM "person" AS "person_1"
LEFT JOIN (
  SELECT "visit_occurrence_1"."person_id", MAX("visit_occurrence_1"."visit_start_date") AS "max"
  FROM "visit_occurrence" AS "visit_occurrence_1"
  GROUP BY "visit_occurrence_1"."person_id"
) AS "visit_group_1" ON ("person_1"."person_id" = "visit_group_1"."person_id")

FunSQL.Append — Method

Append(; over = nothing, list)
Append(list...; over = nothing)

A subquery that combines subqueries using UNION ALL.

SELECT ...
FROM $over
UNION ALL
SELECT ...
FROM $(list[1])
UNION ALL
...

Examples

julia> measurement = SQLTable(:measurement, columns = [:measurement_id, :person_id, :measurement_date]);

julia> observation = SQLTable(:observation, columns = [:observation_id, :person_id, :observation_date]);

julia> q = From(measurement) |>
           Define(:date => Get.measurement_date) |>
           Append(From(observation) |>
                  Define(:date => Get.observation_date)) |>
           Select(Get.person_id, Get.date);

julia> print(render(q))
SELECT "union_1"."person_id", "union_1"."date"
FROM (
  SELECT "measurement_1"."person_id", "measurement_1"."measurement_date" AS "date"
  FROM "measurement" AS "measurement_1"
  UNION ALL
  SELECT "observation_1"."person_id", "observation_1"."observation_date" AS "date"
  FROM "observation" AS "observation_1"
) AS "union_1"

FunSQL.As — Method

As(; over = nothing, name)
As(name; over = nothing)
name => over

An alias for a subquery or an expression.

Examples

julia> person = SQLTable(:person, columns = [:person_id, :year_of_birth]);

julia> q = From(person) |>
           As(:p) |>
           Select(:birth_year => Get.p.year_of_birth);

FunSQL.Bind — Method

Bind(; over = nothing; list)
Bind(list...; over = nothing)

Bind a query parameter to make a correlated subquery.

Examples

julia> person = SQLTable(:person, columns = [:person_id]);

julia> visit_occurrence = SQLTable(:visit_occurrence, columns = [:visit_occurrence_id, :person_id]);

julia> q = From(person) |>
           Where(Fun.exists(From(visit_occurrence) |>
                            Where(Get.person_id .== Var.person_id) |>
                            Bind(Get.person_id)));

julia> print(render(q))
SELECT "person_1"."person_id"
FROM "person" AS "person_1"
WHERE (EXISTS (
  SELECT NULL
  FROM "visit_occurrence" AS "visit_occurrence_1"
  WHERE ("visit_occurrence_1"."person_id" = "person_1"."person_id")
))

FunSQL.Define — Method

Define(; over; list = [])
Define(list...; over)

A subquery that defines calculated columns.

Examples

julia> person = SQLTable(:person, columns = [:person_id, :birth_datetime]);

julia> q = From(person) |>
           Define(:age => Fun.now() .- Get.birth_datetime) |>
           Where(Get.age .> "16 years") |>
           Select(Get.person_id, Get.age);

julia> print(render(q))
SELECT "person_1"."person_id", (NOW() - "person_1"."birth_datetime") AS "age"
FROM "person" AS "person_1"
WHERE ((NOW() - "person_1"."birth_datetime") > '16 years')

FunSQL.From — Method

From(; table)
From(table)

A subquery that selects columns from the given table.

SELECT ...
FROM $table

Examples

julia> person = SQLTable(:person, columns = [:person_id, :year_of_birth]);

julia> q = From(person);

julia> print(render(q))
SELECT "person_1"."person_id", "person_1"."year_of_birth"
FROM "person" AS "person_1"

FunSQL.Fun — Method

Fun(; name, args = [])
Fun(name; args = [])
Fun(name, args...)

A function or an operator invocation.

Example

julia> person = SQLTable(:person, columns = [:person_id, :year_of_birth]);

julia> q = From(person) |>
           Where(Fun.not(Get.person_id .> 2000));

julia> print(render(q))
SELECT "person_1"."person_id", "person_1"."year_of_birth"
FROM "person" AS "person_1"
WHERE (NOT ("person_1"."person_id" > 2000))

FunSQL.Get — Method

Get(; over, name)
Get(name; over)
Get.name        Get."name"      Get[name]       Get["name"]
over.name       over."name"     over[name]      over["name"]

A reference to a table column, or an aliased expression or subquery.

Examples

julia> person = SQLTable(:person, columns = [:person_id, :year_of_birth]);

julia> q = From(person) |>
           As(:p) |>
           Select(Get.p.person_id);

julia> person = SQLTable(:person, columns = [:person_id, :year_of_birth]);

julia> q = From(person);

julia> q = q |> Select(q.person_id);

FunSQL.Group — Method

Group(; over; by = [])
Group(by...; over)

A subquery that groups rows by a list of keys.

SELECT ...
FROM $over
GROUP BY $by...

Examples

julia> person = SQLTable(:person, columns = [:person_id, :year_of_birth]);

julia> q = From(person) |>
           Group(Get.year_of_birth) |>
           Select(Get.year_of_birth, Agg.count());

julia> print(render(q))
SELECT "person_1"."year_of_birth", COUNT(*) AS "count"
FROM "person" AS "person_1"
GROUP BY "person_1"."year_of_birth"

julia> person = SQLTable(:person, columns = [:person_id, :year_of_birth]);

julia> q = From(person) |>
           Group() |>
           Select(Agg.count(distinct = true, Get.year_of_birth));

julia> print(render(q))
SELECT COUNT(DISTINCT "person_1"."year_of_birth") AS "count"
FROM "person" AS "person_1"

FunSQL.Highlight — Method

Highlight(; over = nothing; color)
Highlight(color; over = nothing)

Highlight over with the given color.

Available colors can be found in Base.text_colors.

Examples

julia> q = Get.person_id |> Highlight(:bold);

FunSQL.Join — Method

Join(; over = nothing, joinee, on, left = false, right = false)
Join(joinee; over = nothing, on, left = false, right = false)
Join(joinee, on; over = nothing, left = false, right = false)

A subquery that joins two subqueries together.

SELECT ...
FROM $over
JOIN $joinee ON $on

Examples

julia> person = SQLTable(:person, columns = [:person_id, :location_id]);

julia> location = SQLTable(:location, columns = [:location_id, :state]);

julia> q = From(person) |>
           Join(:location => location,
                Get.location_id .== Get.location.location_id) |>
           Select(Get.person_id, Get.location.state);

julia> print(render(q))
SELECT "person_1"."person_id", "location_1"."state"
FROM "person" AS "person_1"
JOIN "location" AS "location_1" ON ("person_1"."location_id" = "location_1"."location_id")

FunSQL.LeftJoin — Method

An alias for Join(...; ..., left = true).

FunSQL.Limit — Method

Limit(; over = nothing, offset = nothing, limit = nothing)
Limit(limit; over = nothing, offset = nothing)
Limit(offset, limit; over = nothing)
Limit(start:stop; over = nothing)

A subquery that takes a fixed-sized slice of the dataset.

SELECT ...
FROM $over
OFFSET $offset ROWS
FETCH NEXT $limit ROWS ONLY

Examples

julia> person = SQLTable(:person, columns = [:person_id]);

julia> q = From(person) |>
           Limit(1) |>
           Select(Get.person_id);

julia> print(render(q))
SELECT "person_1"."person_id"
FROM "person" AS "person_1"
FETCH FIRST 1 ROW ONLY

FunSQL.Lit — Method

Lit(; val)
Lit(val)

A SQL literal.

In a suitable context, missing, numbers, strings and datetime values are automatically converted to SQL literals.

Examples

julia> q = Select(:null => missing,
                  :boolean => true,
                  :integer => 42,
                  :text => "SQL is fun!",
                  :date => Date(2000));

julia> print(render(q))
SELECT NULL AS "null", TRUE AS "boolean", 42 AS "integer", 'SQL is fun!' AS "text", '2000-01-01' AS "date"

FunSQL.Order — Method

Order(; over = nothing, by)
Order(by...; over = nothing)

A subquery that sorts the rows by a list of keys.

SELECT ...
FROM $over
ORDER BY $by...

Examples

julia> person = SQLTable(:person, columns = [:person_id, :year_of_birth]);

julia> q = From(person) |>
           Order(Get.year_of_birth) |>
           Select(Get.person_id);

julia> print(render(q))
SELECT "person_1"."person_id"
FROM "person" AS "person_1"
ORDER BY "person_1"."year_of_birth"

FunSQL.Partition — Method

Partition(; over, by = [], order_by = [], frame = nothing)
Partition(by...; over, order_by = [], frame = nothing)

A subquery that partitions rows by a list of keys.

SELECT ...
FROM $over
WINDOW w AS (PARTITION BY $by... ORDER BY $order_by...)

Examples

julia> person = SQLTable(:person, columns = [:person_id, :year_of_birth]);

julia> q = From(person) |>
           Partition(Get.year_of_birth) |>
           Select(Get.year_of_birth, Agg.row_number());

julia> print(render(q))
SELECT "person_1"."year_of_birth", (ROW_NUMBER() OVER (PARTITION BY "person_1"."year_of_birth")) AS "row_number"
FROM "person" AS "person_1"

julia> person = SQLTable(:person, columns = [:person_id, :year_of_birth]);

julia> q = From(person) |>
           Group(Get.year_of_birth) |>
           Partition(order_by = [Get.year_of_birth],
                     frame = (mode = :range, start = -1, finish = 1)) |>
           Select(Get.year_of_birth, Agg.avg(Agg.count()));

julia> print(render(q))
SELECT "person_1"."year_of_birth", (AVG(COUNT(*)) OVER (ORDER BY "person_1"."year_of_birth" RANGE BETWEEN 1 PRECEDING AND 1 FOLLOWING)) AS "avg"
FROM "person" AS "person_1"
GROUP BY "person_1"."year_of_birth"

FunSQL.Select — Method

Select(; over; list)
Select(list...; over)

A subquery that fixes the list of output columns.

SELECT $list...
FROM $over

Examples

julia> person = SQLTable(:person, columns = [:person_id, :year_of_birth]);

julia> q = From(person) |>
           Select(Get.person_id);

julia> print(render(q))
SELECT "person_1"."person_id"
FROM "person" AS "person_1"

FunSQL.Asc — Method

Asc(; over = nothing, nulls = nothing)

Ascending order indicator.

FunSQL.Desc — Method

Desc(; over = nothing, nulls = nothing)

Descending order indicator.

FunSQL.Sort — Method

Sort(; over = nothing, value, nulls = nothing)
Sort(value; over = nothing, nulls = nothing)
Asc(; over = nothing, nulls = nothing)
Desc(; over = nothing, nulls = nothing)

Sort order indicator.

Examples

julia> person = SQLTable(:person, columns = [:person_id, :year_of_birth]);

julia> q = From(person) |>
           Order(Get.year_of_birth |> Desc()) |>
           Select(Get.person_id);

julia> print(render(q))
SELECT "person_1"."person_id"
FROM "person" AS "person_1"
ORDER BY "person_1"."year_of_birth" DESC

FunSQL.Var — Method

Var(; name)
Var(name)
Var.name        Var."name"      Var[name]       Var["name"]

A reference to a query parameter.

Examples

julia> person = SQLTable(:person, columns = [:person_id, :year_of_birth]);

julia> q = From(person) |>
           Where(Get.year_of_birth .> Var.year);

julia> print(render(q))
SELECT "person_1"."person_id", "person_1"."year_of_birth"
FROM "person" AS "person_1"
WHERE ("person_1"."year_of_birth" > :year)

FunSQL.Where — Method

Where(; over = nothing, condition)
Where(condition; over = nothing)

A subquery that filters by the given condition.

SELECT ...
FROM $over
WHERE $condition

Examples

julia> person = SQLTable(:person, columns = [:person_id, :year_of_birth]);

julia> q = From(person) |>
           Where(Fun(">", Get.year_of_birth, 2000));

julia> print(render(q))
SELECT "person_1"."person_id", "person_1"."year_of_birth"
FROM "person" AS "person_1"
WHERE ("person_1"."year_of_birth" > 2000)

Syntactic Structure

FunSQL.AbstractSQLClause — Type

A part of a SQL query.

FunSQL.SQLClause — Type

An opaque wrapper over an arbitrary SQL clause.

FunSQL.AGG — Method

AGG(; name, distinct = false, args = [], filter = nothing, over = nothing)
AGG(name; distinct = false, args = [], filter = nothing, over = nothing)
AGG(name, args...; distinct = false, filter = nothing, over = nothing)

An application of an aggregate function.

Examples

julia> c = AGG(:COUNT, OP("*"));

julia> print(render(c))
COUNT(*)

julia> c = AGG(:COUNT, distinct = true, :year_of_birth);

julia> print(render(c))
COUNT(DISTINCT "year_of_birth")

julia> c = AGG(:COUNT, OP("*"), filter = OP(">", :year_of_birth, 1970));

julia> print(render(c))
(COUNT(*) FILTER (WHERE ("year_of_birth" > 1970)))

julia> c = AGG(:ROW_NUMBER, over = PARTITION(:year_of_birth));

julia> print(render(c))
(ROW_NUMBER() OVER (PARTITION BY "year_of_birth"))

FunSQL.AS — Method

AS(; over = nothing, name)
AS(name; over = nothing)

An AS clause.

Examples

julia> c = ID(:person) |> AS(:p);

julia> print(render(c))
"person" AS "p"

FunSQL.CASE — Method

CASE(; args)
CASE(args...)

A CASE expression.

Examples

julia> c = CASE(OP("<", :year_of_birth, 1970), "boomer", "millenial");

julia> print(render(c))
(CASE WHEN ("year_of_birth" < 1970) THEN 'boomer' ELSE 'millenial' END)

FunSQL.FROM — Method

FROM(; over = nothing)
FROM(over)

A FROM clause.

Examples

julia> c = ID(:person) |> AS(:p) |> FROM() |> SELECT((:p, :person_id));

julia> print(render(c))
SELECT "p"."person_id"
FROM "person" AS "p"

FunSQL.FUN — Method

FUN(; name, args = [])
FUN(name; args = [])
FUN(name, args...)

An invocation of a SQL function.

Examples

julia> c = FUN(:EXTRACT, OP(:YEAR), KW(:FROM, FUN(:NOW)));

julia> print(render(c))
EXTRACT(YEAR FROM NOW())

FunSQL.GROUP — Method

GROUP(; over = nothing, by = [])
GROUP(by...; over = nothing)

A GROUP BY clause.

Examples

julia> c = FROM(:person) |>
           GROUP(:year_of_birth) |>
           SELECT(:year_of_birth, AGG("COUNT", OP("*")));

julia> print(render(c))
SELECT "year_of_birth", COUNT(*)
FROM "person"
GROUP BY "year_of_birth"

FunSQL.HAVING — Method

HAVING(; over = nothing, condition)
HAVING(condition; over = nothing)

A HAVING clause.

Examples

julia> c = FROM(:person) |>
           GROUP(:year_of_birth) |>
           HAVING(OP(">", AGG("COUNT", OP("*")), 10)) |>
           SELECT(:person_id);

julia> print(render(c))
SELECT "person_id"
FROM "person"
GROUP BY "year_of_birth"
HAVING (COUNT(*) > 10)

FunSQL.ID — Method

ID(; over = nothing, name)
ID(name; over = nothing)

A SQL identifier. Specify over or use the |> operator to make a qualified identifier.

Examples

julia> c = ID(:person);

julia> print(render(c))
"person"

julia> c = ID(:p) |> ID(:birth_datetime);

julia> print(render(c))
"p"."birth_datetime"

FunSQL.JOIN — Method

JOIN(; over = nothing, joinee, on, left = false, right = false, lateral = false)
JOIN(joinee; over = nothing, on, left = false, right = false, lateral = false)
JOIN(joinee, on; over = nothing, left = false, right = false, lateral = false)

A JOIN clause.

Examples

julia> c = FROM(:p => :person) |>
           JOIN(:l => :location,
                on = OP("=", (:p, :location_id), (:l, :location_id)),
                left = true) |>
           SELECT((:p, :person_id), (:l, :state));

julia> print(render(c))
SELECT "p"."person_id", "l"."state"
FROM "person" AS "p"
LEFT JOIN "location" AS "l" ON ("p"."location_id" = "l"."location_id")

FunSQL.KW — Method

KW(; over = nothing, name)
KW(name; over = nothing)
KW(over, name)

A keyword argument of a function or an operator.

Examples

julia> c = FUN(:SUBSTRING, :zip, KW(:FROM, 1), KW(:FOR, 3));

julia> print(render(c))
SUBSTRING("zip" FROM 1 FOR 3)

julia> c = OP(:BETWEEN, :year_of_birth, 2000, KW(:AND, 2010));

julia> print(render(c))
("year_of_birth" BETWEEN 2000 AND 2010)

FunSQL.LIMIT — Method

LIMIT(; over = nothing, offset = nothing, limit = nothing, with_ties = false)
LIMIT(limit; over = nothing, offset = nothing, with_ties = false)
LIMIT(offset, limit; over = nothing, with_ties = false)
LIMIT(start:stop; over = nothing, with_ties = false)

A LIMIT clause.

Examples

julia> c = FROM(:person) |>
           LIMIT(1) |>
           SELECT(:person_id);

julia> print(render(c))
SELECT "person_id"
FROM "person"
FETCH FIRST 1 ROW ONLY

FunSQL.LIT — Method

LIT(; val)
LIT(val)

A SQL literal.

In a context of a SQL clause, missing, numbers, strings and datetime values are automatically converted to SQL literals.

Examples

julia> c = LIT(missing);

julia> print(render(c))
NULL

julia> c = LIT("SQL is fun!");

julia> print(render(c))
'SQL is fun!'

FunSQL.OP — Method

OP(; name, args = [])
OP(name; args = [])
OP(name, args...)

An application of a SQL operator.

Examples

julia> c = OP("NOT", OP("=", :zip, "60614"));

julia> print(render(c))
(NOT ("zip" = '60614'))

FunSQL.ORDER — Method

ORDER(; over = nothing, by = [])
ORDER(by...; over = nothing)

A ORDER BY clause.

Examples

julia> c = FROM(:person) |>
           ORDER(:year_of_birth) |>
           SELECT(:person_id);

julia> print(render(c))
SELECT "person_id"
FROM "person"
ORDER BY "year_of_birth"

FunSQL.PARTITION — Method

PARTITION(; over = nothing, by = [], order_by = [], frame = nothing)
PARTITION(by...; over = nothing, order_by = [], frame = nothing)

A window definition clause.

Examples

julia> c = FROM(:person) |>
           SELECT(:person_id,
                  AGG("ROW_NUMBER", over = PARTITION(:year_of_birth)));

julia> print(render(c))
SELECT "person_id", (ROW_NUMBER() OVER (PARTITION BY "year_of_birth"))
FROM "person"

julia> c = FROM(:person) |>
           WINDOW(:w1 => PARTITION(:year_of_birth),
                  :w2 => :w1 |> PARTITION(order_by = [:month_of_birth, :day_of_birth])) |>
           SELECT(:person_id, AGG("ROW_NUMBER", over = :w2));

julia> print(render(c))
SELECT "person_id", (ROW_NUMBER() OVER ("w2"))
FROM "person"
WINDOW "w1" AS (PARTITION BY "year_of_birth"), "w2" AS ("w1" ORDER BY "month_of_birth", "day_of_birth")

julia> c = FROM(:person) |>
           GROUP(:year_of_birth) |>
           SELECT(:year_of_birth,
                  AGG("AVG",
                      AGG("COUNT", OP("*")),
                      over = PARTITION(order_by = [:year_of_birth],
                                       frame = (mode = :range, start = -1, finish = 1))));

julia> print(render(c))
SELECT "year_of_birth", (AVG(COUNT(*)) OVER (ORDER BY "year_of_birth" RANGE BETWEEN 1 PRECEDING AND 1 FOLLOWING))
FROM "person"
GROUP BY "year_of_birth"

FunSQL.SELECT — Method

SELECT(; over = nothing, top = nothing, distinct = false, list)
SELECT(list...; over = nothing, top = nothing, distinct = false)

A SELECT clause. Unlike raw SQL, SELECT() should be placed at the end of a clause chain.

Set distinct to true to add a DISTINCT modifier.

Examples

julia> c = SELECT(true, false);

julia> print(render(c))
SELECT TRUE, FALSE

julia> c = FROM(:location) |>
           SELECT(distinct = true, :zip);

julia> print(render(c))
SELECT DISTINCT "zip"
FROM "location"

FunSQL.ASC — Method

ASC(; over = nothing, nulls = nothing)

Ascending order indicator.

FunSQL.DESC — Method

DESC(; over = nothing, nulls = nothing)

Descending order indicator.

FunSQL.SORT — Method

SORT(; over = nothing, value, nulls = nothing)
SORT(value; over = nothing, nulls = nothing)
ASC(; over = nothing, nulls = nothing)
DESC(; over = nothing, nulls = nothing)

Sort order options.

Examples

julia> c = FROM(:person) |>
           ORDER(:year_of_birth |> DESC()) |>
           SELECT(:person_id);

julia> print(render(c))
SELECT "person_id"
FROM "person"
ORDER BY "year_of_birth" DESC

FunSQL.UNION — Method

UNION(; over = nothing, all = false, list)
UNION(list...; over = nothing, all = false)

A UNION clause.

Examples

julia> c = FROM(:measurement) |>
           SELECT(:person_id, :date => :measurement_date) |>
           UNION(all = true,
                 FROM(:observation) |>
                 SELECT(:person_id, :date => :observation_date));

julia> print(render(c))
SELECT "person_id", "measurement_date" AS "date"
FROM "measurement"
UNION ALL
SELECT "person_id", "observation_date" AS "date"
FROM "observation"

FunSQL.VAR — Method

VAR(; name)
VAR(name)

A placeholder parameter in a prepared statement.

Examples

julia> c = VAR(:year);

julia> print(render(c))
:year

FunSQL.WHERE — Method

WHERE(; over = nothing, condition)
WHERE(condition; over = nothing)

A WHERE clause.

Examples

julia> c = FROM(:location) |>
           WHERE(OP("=", :zip, "60614")) |>
           SELECT(:location_id);

julia> print(render(c))
SELECT "location_id"
FROM "location"
WHERE ("zip" = '60614')

FunSQL.WINDOW — Method

WINDOW(; over = nothing, list)
WINDOW(list...; over = nothing)

A WINDOW clause.

Examples

julia> c = FROM(:person) |>
           WINDOW(:w1 => PARTITION(:year_of_birth),
                  :w2 => :w1 |> PARTITION(order_by = [:month_of_birth, :day_of_birth])) |>
           SELECT(:person_id, AGG("ROW_NUMBER", over = :w2));

julia> print(render(c))
SELECT "person_id", (ROW_NUMBER() OVER ("w2"))
FROM "person"
WINDOW "w1" AS (PARTITION BY "year_of_birth"), "w2" AS ("w1" ORDER BY "month_of_birth", "day_of_birth")