FunSQL.jl

API Reference

FunSQL.render — Function

render(node; dialect = :default) :: String

Convert the given SQL node or clause object to a SQL string.

SQL Dialects

FunSQL.SQLDialect — Type

Properties of a SQL dialect.

SQL Entities

FunSQL.SQLTable — Type

SQLTable(; schema = nothing, name, columns)
SQLTable(name; schema = nothing, columns)
SQLTable(name, columns...; schema = nothing)

The structure of a SQL table or a table-like entity (TEMP TABLE, VIEW, etc) for use as a reference in assembling SQL queries.

The SQLTable constructor expects the table name, a vector columns of column names, and, optionally, the name of the table schema. A name can be provided as a Symbol or String value.

Examples

julia> t = SQLTable(:location,
                    :location_id, :address_1, :address_2, :city, :state, :zip);

julia> show(t.name)
:location

julia> show(t.columns)
[:location_id, :address_1, :address_2, :city, :state, :zip]

julia> t = SQLTable(schema = "public",
                    name = "person",
                    columns = ["person_id", "birth_datetime", "location_id"]);

julia> show(t.schema)
:public

julia> show(t.name)
:person

julia> show(t.columns)
[:person_id, :birth_datetime, :location_id]

Semantic Structure

FunSQL.AbstractSQLNode — Type

A SQL operation.

FunSQL.SQLNode — Type

An opaque wrapper over an arbitrary SQL node.

FunSQL.resolve — Method

resolve(node; dialect = :default)

Convert the node to a SELECT clause.

FunSQL.As — Method

As(; over = nothing; name)
As(name; over = nothing)
name => over

An alias for a subquery or an expression.

Examples

julia> person = SQLTable(:person, columns = [:person_id, :year_of_birth]);

julia> q = From(person) |>
           As(:p) |>
           Select(:birth_year => Get.p.year_of_birth);

FunSQL.Call — Method

Call(; name, args = [])
Call(name; args = [])
Call(name, args...)

A function or an operator invocation.

Example

julia> person = SQLTable(:person, columns = [:person_id, :year_of_birth]);

julia> q = From(person) |>
           Where(Call("NOT", Call(">", Get.person_id, 2000)));

julia> print(render(q))
SELECT "person_2"."person_id", "person_2"."year_of_birth"
FROM (
  SELECT "person_1"."person_id", "person_1"."year_of_birth"
  FROM "person" AS "person_1"
) AS "person_2"
WHERE (NOT ("person_2"."person_id" > 2000))

FunSQL.From — Method

From(; table)
From(table)

A subquery that selects columns from the given table.

SELECT ... FROM $table

Examples

julia> person = SQLTable(:person, columns = [:person_id, :year_of_birth]);

julia> q = From(person);

julia> print(render(q))
SELECT "person_1"."person_id", "person_1"."year_of_birth"
FROM "person" AS "person_1"

FunSQL.Get — Method

Get(; over, name)
Get(name; over)
Get.name        Get."name"      Get[name]       Get["name"]
over.name       over."name"     over[name]      over["name"]

A reference to a table column, or an aliased expression or subquery.

Examples

julia> person = SQLTable(:person, columns = [:person_id, :year_of_birth]);

julia> q = From(person) |>
           As(:p) |>
           Select(Get.p.person_id);

julia> person = SQLTable(:person, columns = [:person_id, :year_of_birth]);

julia> q = From(person);

julia> q = q |> Select(q.person_id);

FunSQL.Highlight — Method

Highlight(; over = nothing; color)
Highlight(color; over = nothing)

Highlight over with the given color.

Available colors can be found in Base.text_colors.

Examples

julia> q = Get.person_id |> Highlight(:bold);

FunSQL.Literal — Method

Literal(val)

A SQL literal.

In a suitable context, missing, numbers, strings and datetime values are automatically converted to SQL literals.

Examples

julia> q = Select(:null => missing,
                  :boolean => true,
                  :integer => 42,
                  :text => "SQL is fun!",
                  :date => Date(2000));

julia> print(render(q))
SELECT NULL AS "null", TRUE AS "boolean", 42 AS "integer", 'SQL is fun!' AS "text", '2000-01-01' AS "date"
FROM (
  SELECT TRUE
) AS "__1"

FunSQL.Select — Method

Select(; over; list)
Select(list...; over)

A subquery that fixes the list of output columns.

SELECT $list... FROM $over

Examples

julia> person = SQLTable(:person, columns = [:person_id, :year_of_birth]);

julia> q = From(person) |>
           Select(Get.person_id);

julia> print(render(q))
SELECT "person_2"."person_id"
FROM (
  SELECT "person_1"."person_id"
  FROM "person" AS "person_1"
) AS "person_2"

FunSQL.Where — Method

Where(; over = nothing, condition)
Where(condition; over = nothing)

A subquery that filters by the given condition.

SELECT ... FROM $over WHERE $condition

Examples

julia> person = SQLTable(:person, columns = [:person_id, :year_of_birth]);

julia> q = From(person) |>
           Where(Call(">", Get.year_of_birth, 2000));

julia> print(render(q))
SELECT "person_2"."person_id", "person_2"."year_of_birth"
FROM (
  SELECT "person_1"."person_id", "person_1"."year_of_birth"
  FROM "person" AS "person_1"
) AS "person_2"
WHERE ("person_2"."year_of_birth" > 2000)

Syntactic Structure

FunSQL.AbstractSQLClause — Type

A part of a SQL query.

FunSQL.SQLClause — Type

An opaque wrapper over an arbitrary SQL clause.

FunSQL.AS — Method

AS(; over = nothing, name)
AS(name; over = nothing)

An AS clause.

Examples

julia> c = ID(:person) |> AS(:p);

julia> print(render(c))
"person" AS "p"

FunSQL.FROM — Method

FROM(; over = nothing)
FROM(over)

A FROM clause.

Examples

julia> c = ID(:person) |> AS(:p) |> FROM() |> SELECT((:p, :person_id));

julia> print(render(c))
SELECT "p"."person_id"
FROM "person" AS "p"

FunSQL.ID — Method

ID(; over = nothing, name)
ID(name; over = nothing)

A SQL identifier. Specify over or use the |> operator to make a qualified identifier.

Examples

julia> c = ID(:person);

julia> print(render(c))
"person"

julia> c = ID(:p) |> ID(:birth_datetime);

julia> print(render(c))
"p"."birth_datetime"

FunSQL.LIT — Method

LIT(val)

A SQL literal.

In a context of a SQL clause, missing, numbers, strings and datetime values are automatically converted to SQL literals.

Examples

julia> c = LIT(missing);

julia> print(render(c))
NULL

julia> c = LIT("SQL is fun!");

julia> print(render(c))
'SQL is fun!'

FunSQL.OP — Method

OP(; name, args)
OP(name; args)
OP(name, args...)

An application of a SQL operator.

Examples

julia> c = OP("NOT", OP("=", :zip, "60614"));

julia> print(render(c))
(NOT ("zip" = '60614'))

FunSQL.SELECT — Method

SELECT(; over = nothing, distinct = false, list)
SELECT(list...; over = nothing, distinct = false)

A SELECT clause. Unlike raw SQL, SELECT() should be placed at the end of a clause chain.

Set distinct to true to add a DISTINCT modifier.

Examples

julia> c = SELECT(true, false);

julia> print(render(c))
SELECT TRUE, FALSE

julia> c = FROM(:location) |>
           SELECT(distinct = true, :zip);

julia> print(render(c))
SELECT DISTINCT "zip"
FROM "location"

FunSQL.WHERE — Method

WHERE(; over = nothing, condition)
WHERE(condition; over = nothing)

A WHERE clause.

Examples

julia> c = FROM(:location) |>
           WHERE(OP("=", :zip, "60614")) |>
           SELECT(:location_id);

julia> print(render(c))
SELECT "location_id"
FROM "location"
WHERE ("zip" = '60614')